Fix: NesEmulator did not mount RAM to CPU bus
