+++ src/devices/mod.rs
@@ -1,5 +1,6 @@
+pub mod bus;
 pub mod cartridge;
 pub mod cpu;
 pub mod nes;
 pub mod ppu;
-pub mod bus;
+pub mod ram;
